AI Powered Tools server/creator-platform/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
Pillow==10.1.0